edit

+++ admin/views/kuliner/tampil.php
@@ -1,8 +1,19 @@
 <script src="https://kit.fontawesome.com/c333d3440e.js" crossorigin="anonymous"></script>
+
+
+<style>
+    .all{
+        padding-top: 10px;
+        padding-left: 50px;
+        padding-right: 50px;
+    }
+   
+</style>
+<div class="all">
 <div class="row">
     <div class="pull-left">
         <br>
-        <h4>Daftar Kuliner</h4>
+        <h4 class="daftar">Daftar Kuliner</h4>
     </div>
     <br>
     <div class="pull-right">
@@ -27,12 +38,14 @@
                     <td><?=$no;?></td><td><?=$row['Jenis_Kuliner'] ?></td><td><?=$row['Nama_Kuliner'];?></td>
                     <td>
                         <a href="index.php?mod=kuliner&page=edit&id=<?=($row["id"])?>"><i class="fa fa-pencil"></i></a>
-                        <a href="index.php?mod=kuliner&page=delete&id=<?=($row["id"])?>"><i class="fa fa-trash"></i></a>
+                        <a href="index.php?mod=kuliner&page=delete&id=<?=($row["id"])?>" onclick="return confirm('Apakah Anda yakin menghapus data ini?')"><i class="fa fa-trash"></i></a>
                    <!--untuk menggunakan font awesome maka run file harus online  -->
+                   
                     </td>
                 </tr>
            <?php $no++; }
             }?>
         </tbody>
     </table>
+</div>
 </div>
